feat: publish src/ folder instead of root
This will allow imports like: `type-level-schema/schema` or
`type-level-schema/validation`.
